[logic][m]: models for search moved to logic - where it actually belongs - refs #328
